refactor: tools내 js파일 기능별 분리
