+++ static/js/tools/getJSON.js
@@ -1,0 +1,3 @@
+export const getJSON = (path) => {
+    return fetch(`/${path}`).then((response) => response.json());
+};
