+++ static/js/tools/preventHyperLinks.js
@@ -1,0 +1,8 @@
+import { $ } from "../utils/dom.js";
+
+export const preventHyperLinks = () => {
+    $("body").addEventListener("click", (e) => {
+        if (!e.target.matches("[data-link]")) return;
+        e.preventDefault();
+    });
+};
