+++ static/js/tools/renderer.js
@@ -1,0 +1,5 @@
+import { $ } from "../utils/dom.js";
+
+export const render = (html) => {
+    $("#app").innerHTML = html;
+};
