+++ static/js/tools/toggleClass.js
@@ -1,0 +1,14 @@
+const toggleClass = (curEl, className) => {
+    const parentNode = curEl.parentNode;
+    Array.from(parentNode.children).forEach((el) =>
+        el.classList.remove(className)
+    );
+    curEl.classList.add(className);
+};
+
+const toggleWeekDayMenu = (target) => {
+    const curEl = target.parentNode;
+    toggleClass(curEl, "active");
+};
+
+export { toggleClass, toggleWeekDayMenu };
